Quick fix and other things
