+++ cvo/CfgFunctions.hpp
@@ -1,0 +1,2 @@
+#include "cbrn\functions.hpp"
+#include "logistics\functions.hpp"
