+++ cvo/CfgSounds.hpp
@@ -1,0 +1,1 @@
+#include "cbrn\sounds.hpp"
