+++ cvo/cbrn/functions.hpp
@@ -1,0 +1,9 @@
+class CVO_CBRN {
+	tag = "cvo_cbrn";
+
+	class EventHandlers {
+		file = "cvo\cbrn\fnc";
+
+		class eventHandler_detector;
+	};
+};
