+++ cvo/cbrn/sounds.hpp
@@ -1,0 +1,17 @@
+class chem_detector_beep_1 {
+    name = "Chemical detector beep 1";
+    sound[] = {"cvo\cbrn\sounds\chem_detector_beep_1.ogg", 5, 1, 20};
+    titles[] = {};
+};
+
+class chem_detector_beep_2 {
+    name = "Chemical detector beep 2";
+    sound[] = {"cvo\cbrn\sounds\chem_detector_beep_2.ogg", 5, 1, 20};
+    titles[] = {};
+};
+
+class chem_detector_beep_3 {
+    name = "Chemical detector beep 3";
+    sound[] = {"cvo\cbrn\sounds\chem_detector_beep_3.ogg", 5, 1, 20};
+    titles[] = {};
+};
