--- cvo/fnc/functions.hpp
@@ -1,20 +1,0 @@
-class CVO_Logistics {
-	tag = "cvo_logistics";
-
-	class Logistics {
-		file = "cvo\fnc\logistics";
-
-		class initCreatedObject;
-		class loadInCargo;
-	};
-};
-
-class CVO_CBRN {
-	tag = "cvo_cbrn";
-
-	class EventHandlers {
-		file = "cvo\fnc\cbrn";
-
-		class eventHandler_detector;
-	};
-};
