+++ cvo/logistics/fn_initCreatedObject.sqf
