+++ cvo/logistics/fn_loadInCargo.sqf
