+++ cvo/logistics/functions.hpp
@@ -1,0 +1,10 @@
+class CVO_Logistics {
+	tag = "cvo_logistics";
+
+	class Logistics {
+		file = "cvo\logistics";
+
+		class initCreatedObject;
+		class loadInCargo;
+	};
+};
