--- cvo/sounds/sounds.hpp
@@ -1,17 +1,0 @@
-class cvo_detector_beep_1 {
-    name = "Chemical detector beep 1";
-    sound[] = {"cvo\sounds\chem_detector_beep_1.ogg", 5, 1, 20};
-    titles[] = {};
-};
-
-class cvo_detector_beep_2 {
-    name = "Chemical detector beep 2";
-    sound[] = {"cvo\sounds\chem_detector_beep_2.ogg", 5, 1, 20};
-    titles[] = {};
-};
-
-class cvo_detector_beep_3 {
-    name = "Chemical detector beep 3";
-    sound[] = {"cvo\sounds\chem_detector_beep_3.ogg", 5, 1, 20};
-    titles[] = {};
-};
